requirements.txt
streamlit==1.45.1
langchain==0.3.25
langchain-community==0.3.24
langchain-google-genai==2.1.4
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.18
python-dotenv==1.1.0
pypdf==5.5.0
faiss-cpu==1.11.0 